requirements.txt
python-binance~=1.0.0
openpyxl~=3.1.0